[UPD] Update resources and add icon

+++ converter/package.ps1
@@ -1,4 +1,13 @@
-rm -Force -Recurse .\build
-rm -Force .\hexconverter.spec
-rm -Force -Recurse '../gui/pyconverter'
+if (Test-Path '.\build') {
+    rm -Force -Recurse '.\build'
+}
+
+if (Test-Path '.\hexconverter.spec') {
+    rm -Force '.\hexconverter.spec'
+}
+
+if (Test-Path '../gui/pyconverter') {
+    rm -Force -Recurse '../gui/pyconverter'
+}
+
 pyinstaller .\hexconverter.py -F --distpath '../gui/pyconverter'
